requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.5.1
branca==0.4.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
folium==0.12.1
gunicorn==20.0.4
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
numpy==1.20.1
pipenv==2020.11.15
psycopg2-binary==2.8.6
pylint==2.7.2
python-telegram-bot==13.4.1
pytz==2021.1
requests==2.25.1
six==1.15.0
toml==0.10.2
tornado==6.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
virtualenv==20.4.3
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wrapt==1.12.1
